requirements.txt
flask
docker-py
pyyaml
mock
